requirements.txt
flask
flask-limiter
Flask-Caching
waitress
python-dotenv
requests